refactor(A6): HomeScreen/AppDrawer dekomponiert
- HomeScreen.kt (1.632 -> ~1.060 Zeilen): Widget-Composables (ClockText/
  DateText/WeatherRow + App-Starts + expandNotifications) nach HomeWidgets.kt,
  FavoriteItem/EditControlButton nach FavoriteItem.kt (gleiches Paket, keine
  Aufrufer-Aenderungen)
- ui/HomeDragStateHolder.kt: buendelt die Edit-Modus-Zustaende (Live-Offsets,
  Neutral-Bounds, Kollisions-/Haptik-Buchfuehrung, Auswahl) mit seedFrom/
  toHomeLayout-Roundtrip; HomeScreen delegiert per Property-Delegation
- AppDrawer: 3x3-Ordner-Grid-Mathematik (Seitenanzahl, Touch->Slot,
  Umsortieren) framework-frei nach LauncherLogic (folderPageCount/
  folderGridSlotAt/moveFolderApp) – die dreifach duplizierte Zell-Rechnung
  ist damit eine getestete Funktion
- Tests: Folder-Grid (Rundung, Zell-Mapping, Clamping, No-Op-Faelle),
  HomeDragStateHolder (Seed/Roundtrip); zwei vorher baselinierte
  MaxLineLength-Stellen echt gefixt
- Kover-Excludes fuer die neuen reinen UI-Dateien

Damit ist Workstream A (Refactoring) des Verbesserungsplans komplett.
Naechster Schritt: B1 (AppWidgetHost) kann auf HomeDragStateHolder aufsetzen.
